admin, announce, hostuserreview api 수정 및 실행테스트

+++ src/main/java/com/example/meetup_study/hostReview/HostReviewServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.meetup_study.hostReview;
 
 import com.example.meetup_study.hostReview.domain.HostReview;
+import com.example.meetup_study.hostReview.domain.dto.HostReviewDto;
 import com.example.meetup_study.hostReview.domain.dto.RequestHostReviewDto;
 import com.example.meetup_study.hostReview.domain.repository.HostReviewRepository;
 import com.example.meetup_study.hostReview.exception.HostReviewInvalidRequestException;
@@ -15,8 +16,10 @@ import com.example.meetup_study.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -27,33 +30,50 @@ public class HostReviewServiceImpl implements HostReviewService{
     private final RoomRepository roomRepository;
     private final UserRepository userRepository;
 
+    @Transactional
     @Override
-    public Optional<HostReview> createHostReview(RequestHostReviewDto requestHostReviewDto, Long userId) {
+    public Optional<HostReviewDto> createHostReview(RequestHostReviewDto requestHostReviewDto, Long userId) {
         Optional<User> userOpt = userRepository.findById(userId);
         Optional<Room> roomOpt = roomRepository.findById(requestHostReviewDto.getRoomId());
         Optional<Review> reviewOpt = reviewRepository.findById(requestHostReviewDto.getReviewId());
 
+        if(reviewOpt.isPresent() && (reviewOpt.get().getIsHostReview() == true)){
+            throw new HostReviewInvalidRequestException("이미 답글 달았습니다");
+        }
+
         HostReview hostReview = hostReviewRepository.save(new HostReview(userOpt.get(), roomOpt.get(), requestHostReviewDto.getContent(), requestHostReviewDto.getReviewId()));
 
-        return Optional.ofNullable(hostReview);
+        reviewOpt.get().changeIsHostReview(true);
+
+
+
+        HostReviewDto hostReviewDto = new HostReviewDto().convertToHostReviewDto(hostReview);
+
+        return Optional.ofNullable(hostReviewDto);
     }
 
     @Override
-    public List<HostReview> findByRoomId(Long roomId) {
+    public List<HostReviewDto> findByRoomId(Long roomId) {
         Optional<Room> room = roomRepository.findById(roomId);
         if(room.isPresent()) {
 
             List<HostReview> hostReviews = hostReviewRepository.findByRoomId(roomId);
 
-            return hostReviews;
+            List<HostReviewDto> hostReviewDtoList = hostReviews.stream()
+                    .map(hostReview -> new HostReviewDto().convertToHostReviewDto(hostReview))
+                    .collect(Collectors.toList());
+
+            return hostReviewDtoList;
 
         }else{
             throw new RoomNotFoundException();
         }
     }
 
+    @Transactional
     @Override
-    public Optional<HostReview> deleteHostReview(Long hostReviewId, Long userId) {
+    public Optional<HostReviewDto> deleteHostReview(Long hostReviewId, Long userId) {
+
         Optional<HostReview> hostReviewOpt = hostReviewRepository.findById(hostReviewId);
 
         if(!hostReviewOpt.isPresent()){
@@ -64,7 +84,16 @@ public class HostReviewServiceImpl implements HostReviewService{
             throw new HostReviewInvalidRequestException();
         }
 
+        Optional<Review> reviewOpt = reviewRepository.findById(hostReviewOpt.get().getReviewId());
+        if (!reviewOpt.isPresent()) {
+            throw new HostReviewNotFoundException();
+        }
+
         hostReviewRepository.deleteById(hostReviewId);
-        return hostReviewOpt;
+        reviewOpt.get().changeIsHostReview(false);
+
+        HostReviewDto hostReviewDto = new HostReviewDto().convertToHostReviewDto(hostReviewOpt.get());
+
+        return Optional.ofNullable(hostReviewDto);
     }
 }
